comfyui_nodes/cineflow_api/requirements.txt
# CineFlow ComfyUI 节点依赖
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
requests>=2.28.0
numpy>=1.21.0
Pillow>=9.0.0
